Rename `CoinSelection.Gen` to `CoinSelection.Balance.Gen`.
The generators and shrinkers in this module are for types defined in the
`Balance` module.
